utils/frame_queue: use doubles for PTS
Instead of floats. Floating precision is around 1e-7f, so after ~10000
seconds of playback (a few hours), rounding error alone will account for
1 ms of jitter in timestamps. This is definitely undesirable.

Switching to double defers this problem until several millions of years
of playback.

It's worth pointing out that we don't need to upgrade all doubles in
this code, just the ones that directly represent absolute-scale
timestamps. All relative values (e.g. for FPS estimates) can remain
float.

+++ src/include/libplacebo/utils/frame_queue.h
@@ -46,7 +46,7 @@ struct pl_source_frame {
     // frame. These must be monotonically increasing for subsequent frames.
     // To implement a discontinuous jump, users must explicitly reset the
     // frame queue with `pl_queue_reset` and restart from PTS 0.0.
-    float pts;
+    double pts;
 
     // The frame's duration. This is not needed in normal scenarios, as the
     // FPS can be inferred from the `pts` values themselves. Providing it
@@ -139,7 +139,7 @@ struct pl_queue_params {
     // These must be monotonically increasing. To implement a discontinuous
     // jump, users must explicitly reset the frame queue with `pl_queue_reset`
     // and restart from PTS 0.0.
-    float pts;
+    double pts;
 
     // The radius of the configured mixer. This should be set to the value
     // as returned by `pl_frame_mix_radius`.
